Refactor to move to appropriate module and package
