+++ build.gradle
@@ -268,6 +268,7 @@ project('test:testcommon') {
         compile group: 'commons-io', name: 'commons-io', version: commonsioVersion
         compile group: 'io.netty', name: 'netty-all', version: nettyVersion
         compile group: 'org.apache.curator', name: 'curator-test', version: apacheCuratorVersion, withoutLogger
+        compile group: 'com.google.code.gson', name: 'gson', version: gsonVersion
     }
     javadoc {
         dependsOn delombok
